Refactor: Move URLBuilder closer to its usages
